[update] accounts/views.py を accounts/views/に移動 & Basic認証FBV -> CBVへ変更
