+++ docker/srcs/uwsgi-django/accounts/templates/accounts/login.html
@@ -8,12 +8,13 @@
 <body>
 <h1>ログイン</h1>
 <p>{{message}}</p>
-<form action="{% url 'login'%}" method="post">
+<form action="{% url 'accounts:login'%}" method="post">
     {% csrf_token %}
     {{ form.as_p }}
     <p><input type="hidden" name="next" value="{{next}}"></p>
     <p><input type="submit" value="ログイン"></p>
 </form>
-<p><a href="{% url 'signup'%}">ユーザー登録</a></p>
+<p><a href="{% url 'accounts:signup'%}">ユーザー登録</a></p>
+<a href="/pong/">🏓 return to pong</a></li>
 </body>
 </html>
