+++ docker/srcs/uwsgi-django/accounts/templates/accounts/logout.html
@@ -6,7 +6,7 @@
     <title>ログアウト</title>
 </head>
 <body>
-<p>ログアウトしました...</p>
+<p>good bye...</p>
 
 <a href="/pong/">🏓 return to pong</a></li>
